Separar responsabilidades del menu y client del patron Factory Method
